Update Game Loop, started loading
Scene descriptions are now programmed in, which are lists of gameobjects. the main load function now checks the current scenes objects, and if they have a texture component, it loads the texture and creates a copy in the games loaded objects. This is a way of loading scenes, and could be expanded later to load and unload during gameplay.