Re-arranged file structure
There should now be no problems with building the .exe, as the Fna.Core .csproj and its files werent included in the github folder. They are now, and so it can be utilized seperately as a build resource for the .exe. This should be fine for a while. Final builds will happen with this, then be shipped with the file structure of the main project folder, and no FNA libraries, as they compile into the .exe? Going to double check but i think they do which is why it is platform agnostic.